refactor: put schema closer to resolvers
